add refactor code
