+++ book_creator.rb
@@ -1,0 +1,19 @@
+require_relative 'book'
+
+class BookCreator
+    def initialize(books)
+      @books = books
+    end
+  
+    def create_book
+      puts 'Title:'
+      title = gets.chomp
+      puts 'Author:'
+      author = gets.chomp
+  
+      book = Book.new(title, author)
+      @books << book
+      puts 'Book created successfully.'
+    end
+  end
+
